+++ apps/rosen/app/(bridge)/layout.tsx
@@ -14,6 +14,8 @@ const Background = styled('div')(({ theme }) => ({
       ? `linear-gradient(180deg,${theme.palette.primary.light},${theme.palette.background.default},${theme.palette.background.default},${theme.palette.secondary.light})`
       : 'none',
   'zIndex': '0',
+  'borderTopLeftRadius': theme.shape.borderRadius * 2,
+  'borderBottomLeftRadius': theme.shape.borderRadius * 2,
   '& > svg': {
     'position': 'absolute',
     'width': '100%',
@@ -36,14 +38,14 @@ const Main = styled('main')(({ theme }) => ({
   'position': 'relative',
   'display': 'grid',
   'gridTemplateRows': 'auto',
-  'gridTemplateColumns': '1fr',
+  'gridTemplateColumns': 'minmax(0, 1fr)',
   'gap': theme.spacing(2),
   'alignContent': 'center',
   'marginInline': 'auto',
   'minHeight': 'calc(100vh - 224px)',
-  'maxWidth': theme.breakpoints.values.laptop,
+  'maxWidth': '1048px',
   [theme.breakpoints.up('laptop')]: {
-    'gridTemplateColumns': '2fr 1fr',
+    'gridTemplateColumns': '2fr minmax(320px, 1fr)',
     'gap': theme.spacing(3),
     '& .alert': {
       gridColumn: '1/-1',