+++ packages/icons/src/index.ts
@@ -1,3 +1,6 @@
+export { ReactComponent as AlignCenter } from './icons/align-center.svg';
+export { default as AlignCenterRaw } from './icons/align-center.svg?raw';
+
 export { ReactComponent as AngleDown } from './icons/angle-down.svg';
 export { default as AngleDownRaw } from './icons/angle-down.svg?raw';
 
@@ -21,6 +24,9 @@ export { default as BitcoinCircleRaw } from './icons/bitcoin-circle.svg?raw';
 
 export { ReactComponent as BitcoinSign } from './icons/bitcoin-sign.svg';
 export { default as BitcoinSignRaw } from './icons/bitcoin-sign.svg?raw';
+
+export { ReactComponent as CaretDown } from './icons/caret-down.svg';
+export { default as CaretDownRaw } from './icons/caret-down.svg?raw';
 
 export { ReactComponent as CheckCircle } from './icons/check-circle.svg';
 export { default as CheckCircleRaw } from './icons/check-circle.svg?raw';
@@ -88,6 +94,9 @@ export { default as KeySkeletonRaw } from './icons/key-skeleton.svg?raw';
 export { ReactComponent as LinkBroken } from './icons/link-broken.svg';
 export { default as LinkBrokenRaw } from './icons/link-broken.svg?raw';
 
+export { ReactComponent as ListUiAlt } from './icons/list-ui-alt.svg';
+export { default as ListUiAltRaw } from './icons/list-ui-alt.svg?raw';
+
 export { ReactComponent as LockAlt } from './icons/lock-alt.svg';
 export { default as LockAltRaw } from './icons/lock-alt.svg?raw';
 
@@ -111,6 +120,9 @@ export { default as PauseRaw } from './icons/pause.svg?raw';
 
 export { ReactComponent as QrcodeScan } from './icons/qrcode-scan.svg';
 export { default as QrcodeScanRaw } from './icons/qrcode-scan.svg?raw';
+
+export { ReactComponent as Qrcode } from './icons/qrcode.svg';
+export { default as QrcodeRaw } from './icons/qrcode.svg?raw';
 
 export { ReactComponent as ReceiptAlt } from './icons/receipt-alt.svg';
 export { default as ReceiptAltRaw } from './icons/receipt-alt.svg?raw';