+++ packages/ui-kit/src/components/common/App.tsx
@@ -55,11 +55,11 @@ const Paper = styled('div', {
 })(({ theme }) => ({
   position: 'relative',
   minHeight: '100vh',
-  overflow: 'hidden',
   backgroundColor: theme.palette.background.default,
   borderTopLeftRadius: theme.shape.borderRadius * 2,
   borderBottomLeftRadius: theme.shape.borderRadius * 2,
   padding: theme.spacing(4),
+  color: theme.palette.text.primary,
   transition: 'ease-in-out 100ms',
   [theme.breakpoints.down('tablet')]: {
     borderTopRightRadius: theme.shape.borderRadius * 2,
@@ -118,7 +118,7 @@ export const App = ({ children, sideBar, theme, toolbar }: AppProps) => {
             ) : (
               sideBar
             )}
-            <Main>
+            <Main ref={main}>
               <Paper>
                 {!isMobile && (
                   <div