+++ packages/ui-kit/src/components/common/display/index.ts
@@ -1,2 +1,4 @@
 export { Amount as Amount2 } from './Amount';
 export * from './Label';
+export * from './Identifier';
+export * from './Network';