+++ packages/ui-kit/src/index.ts
@@ -24,6 +24,8 @@ declare module '@mui/material/styles' {
   }
 }
 
+export * from 'react-sticky-box';
+
 export * from './components';
 export * from './hooks';
 export * from './Providers';