+++ networks/ergo/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @rosen-network/ergo
 
+## 2.1.3
+
+### Patch Changes
+
+- Optimize the structure to enhance compatibility with wallet packages and streamline usage within the Rosen App
+- Update dependencies
+  - @rosen-bridge/icons@2.0.0
+  - @rosen-bridge/tokens@3.1.1
+  - @rosen-clients/ergo-explorer@1.1.5
+  - @rosen-ui/constants@0.4.0
+  - @rosen-ui/wallet-api@2.0.0
+  - @rosen-ui/types@0.3.6
+
 ## 2.1.2
 
 ### Patch Changes